+++ dms-20250414/alibabacloud_dms20250414/models/_add_user_to_data_agent_workspace_response_body.py
@@ -66,8 +66,6 @@ class AddUserToDataAgentWorkspaceResponseBody(DaraModel):
 
         return self
 
-
-
 class AddUserToDataAgentWorkspaceResponseBodyData(DaraModel):
     def __init__(
         self,