+++ dms-20250414/alibabacloud_dms20250414/models/_modify_custom_agent_request.py
@@ -10,6 +10,7 @@ from darabonba.model import DaraModel
 class ModifyCustomAgentRequest(DaraModel):
     def __init__(
         self,
+        callback_config: main_models.ModifyCustomAgentRequestCallbackConfig = None,
         custom_agent_id: str = None,
         dmsunit: str = None,
         data_json: str = None,
@@ -24,6 +25,7 @@ class ModifyCustomAgentRequest(DaraModel):
         web_report_config: str = None,
         workspace_id: str = None,
     ):
+        self.callback_config = callback_config
         # This parameter is required.
         self.custom_agent_id = custom_agent_id
         self.dmsunit = dmsunit
@@ -40,6 +42,8 @@ class ModifyCustomAgentRequest(DaraModel):
         self.workspace_id = workspace_id
 
     def validate(self):
+        if self.callback_config:
+            self.callback_config.validate()
         if self.execution_config:
             self.execution_config.validate()
         if self.knowledge_config_list:
@@ -54,6 +58,9 @@ class ModifyCustomAgentRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.callback_config is not None:
+            result['CallbackConfig'] = self.callback_config.to_map()
+
         if self.custom_agent_id is not None:
             result['CustomAgentId'] = self.custom_agent_id
 
@@ -99,6 +106,10 @@ class ModifyCustomAgentRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('CallbackConfig') is not None:
+            temp_model = main_models.ModifyCustomAgentRequestCallbackConfig()
+            self.callback_config = temp_model.from_map(m.get('CallbackConfig'))
+
         if m.get('CustomAgentId') is not None:
             self.custom_agent_id = m.get('CustomAgentId')
 
@@ -271,6 +282,65 @@ class ModifyCustomAgentRequestExecutionConfig(DaraModel):
 
         if m.get('SkipWebReportConfirm') is not None:
             self.skip_web_report_confirm = m.get('SkipWebReportConfirm')
+
+        return self
+
+class ModifyCustomAgentRequestCallbackConfig(DaraModel):
+    def __init__(
+        self,
+        callback_args: str = None,
+        callback_prompt: str = None,
+        callback_time: int = None,
+        tool_id: str = None,
+        type: str = None,
+    ):
+        self.callback_args = callback_args
+        self.callback_prompt = callback_prompt
+        self.callback_time = callback_time
+        self.tool_id = tool_id
+        self.type = type
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.callback_args is not None:
+            result['CallbackArgs'] = self.callback_args
+
+        if self.callback_prompt is not None:
+            result['CallbackPrompt'] = self.callback_prompt
+
+        if self.callback_time is not None:
+            result['CallbackTime'] = self.callback_time
+
+        if self.tool_id is not None:
+            result['ToolId'] = self.tool_id
+
+        if self.type is not None:
+            result['Type'] = self.type
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('CallbackArgs') is not None:
+            self.callback_args = m.get('CallbackArgs')
+
+        if m.get('CallbackPrompt') is not None:
+            self.callback_prompt = m.get('CallbackPrompt')
+
+        if m.get('CallbackTime') is not None:
+            self.callback_time = m.get('CallbackTime')
+
+        if m.get('ToolId') is not None:
+            self.tool_id = m.get('ToolId')
+
+        if m.get('Type') is not None:
+            self.type = m.get('Type')
 
         return self